[refactor] #44 region 조회 api 분리
